requirements.txt
Appium-Python-Client==3.0.0
pytest==7.4.0
pytest-html==3.2.0
allure-pytest==2.13.2
python-dotenv==1.0.0